pyproject.toml
[project]
name = "musicapp"
version = "0.1.0"
description = "Music application"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.128.0",
    "librosa",
    "numba>=0.60",
    "pydantic>=2.12.5",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "faiss-cpu>=1.13.2",
    "numpy>=2.0",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "pydub>=0.25.1",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
]

